# backend/docs/swagger.yaml
basePath: /api
consumes:
- application/json
definitions:
...
  middlewares.Unauthorized:
    properties:
      error:
        type: string
    type: object
...
  models.DailyTraffic:
    properties:
      date:
        description: 'Format: YYYY-MM-DD'
        type: string
      rx:
        description: in GiB
        type: number
      tx:
        description: in GiB
        type: number
    type: object
...
  models.OcservUserSessionEvent:
    enum:
    - user-agent
    - handshake
    - periodic-stats
    - disconnect
    type: string
    x-enum-varnames:
    - EventUseragent
    - EventHandshake
    - EventPeriodicStats
    - EventDisconnect
  models.OcservUserSessionLog:
    properties:
      created_at:
        type: string
      event:
        allOf:
        - $ref: '#/definitions/models.OcservUserSessionEvent'
        enum:
        - user-agent
        - handshake
        - periodic-stats
        - disconnect
      ip:
        type: string
      message:
        type: string
      username:
        type: string
    required:
    - created_at
    - event
    - message
    - username
    type: object
...
  reports.OcservUserReportResponse:
    properties:
      active:
        type: integer
      deactivated:
        type: integer
      locked:
        type: integer
      online:
        type: integer
    type: object
  reports.SessionLogsResponse:
    properties:
      meta:
        $ref: '#/definitions/request.Meta'
      result:
        items:
          $ref: '#/definitions/models.OcservUserSessionLog'
        type: array
    required:
    - meta
    type: object
...
  repository.TotalBandwidths:
    properties:
      rx:
        type: number
      tx:
        type: number
    required:
    - rx
    - tx
    type: object
  request.ErrorResponse:
    properties:
      error:
        items:
          type: string
        type: array
      message:
        items:
          type: string
        type: array
    required:
    - error
    - message
    type: object
  request.Meta:
    properties:
      page:
        type: integer
      size:
        type: integer
      total_records:
        type: integer
    required:
    - page
    - size
    - total_records
    type: object
  runtime.ConfigResponse:
    properties:
      allow_update:
        type: boolean
      auth_timeout:
        type: integer
      ban_reset_time:
        type: integer
      banner:
        type: string
      cisco_client_compat:
        type: boolean
      cookie_timeout:
        type: integer
      deny_roaming:
        type: boolean
      dns:
        items:
          type: string
        type: array
      dpd:
        type: integer
      dtls_legacy:
        type: boolean
      ipv4_network:
        type: string
      keepalive:
        type: integer
      log_level:
        type: integer
      max_ban_score:
        type: integer
      max_clients:
        type: integer
      max_same_clients:
        type: integer
      min_reauth_time:
        type: integer
      mobile_dpd:
        type: integer
      mtu:
        type: integer
      ping_leases:
        type: boolean
      pre_login_banner:
        type: string
      predictable_ips:
        type: boolean
      rate_limit_ms:
        type: integer
      rekey_method:
        allOf:
        - $ref: '#/definitions/system.RekeyMethod'
        enum:
        - ssl
        - new-tunnel
      rekey_time:
        type: integer
      switch_to_tcp_timeout:
        type: integer
      tcp_port:
        type: integer
      try_mtu_discovery:
        type: boolean
      tunnel_all_dns:
        type: boolean
      udp_port:
        type: integer
    type: object
  runtime.OcservConfig:
    properties:
      auth_timeout:
        type: integer
      ban_reset_time:
        type: integer
      banner:
        type: string
      cisco_client_compat:
        type: boolean
      cookie_timeout:
        type: integer
      deny_roaming:
        type: boolean
      dns:
        items:
          type: string
        type: array
      dpd:
        type: integer
      dtls_legacy:
        type: boolean
      ipv4_network:
        type: string
      keepalive:
        type: integer
      log_level:
        type: integer
      max_ban_score:
        type: integer
      max_clients:
        type: integer
      max_same_clients:
        type: integer
      min_reauth_time:
        type: integer
      mobile_dpd:
        type: integer
      mtu:
        type: integer
      ping_leases:
        type: boolean
      pre_login_banner:
        type: string
      predictable_ips:
        type: boolean
      rate_limit_ms:
        type: integer
      rekey_method:
        allOf:
        - $ref: '#/definitions/system.RekeyMethod'
        enum:
        - ssl
        - new-tunnel
      rekey_time:
        type: integer
      switch_to_tcp_timeout:
        type: integer
      tcp_port:
        type: integer
      try_mtu_discovery:
        type: boolean
      tunnel_all_dns:
        type: boolean
      udp_port:
        type: integer
    type: object
...
  system.RekeyMethod:
    enum:
    - ssl
    - new-tunnel
    type: string
    x-enum-varnames:
    - RekeyMethodSSL
    - RekeyMethodNewTunnel
info:
  contact: {}
  description: Administrative API for managing Ocserv Dashboard users, groups, reports,
    runtime, backups, system settings, Telegram integrations, and agent nodes.
  title: Ocserv Dashboard Admin API
  version: "1.0"
paths:
...
  /reports/session_logs:
    get:
      consumes:
      - application/json
      description: Ocserv session logs
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Page number, starting from 1
        in: query
        minimum: 1
        name: page
        type: integer
      - description: Number of items per page
        in: query
        maximum: 100
        minimum: 1
        name: size
        type: integer
      - description: Field to order by
        in: query
        name: order
        type: string
      - description: Sort order, either ASC or DESC
        enum:
        - ASC
        - DESC
        in: query
        name: sort
        type: string
      - description: date_start
        in: query
        name: date_start
        type: string
      - description: date_end
        in: query
        name: date_end
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reports.SessionLogsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: Ocserv session logs
      tags:
      - Report
  /reports/statistics:
    get:
      consumes:
      - application/json
      description: Ocserv Users Statistics
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: date_start
        in: query
        name: date_start
        required: true
        type: string
      - description: date_end
        in: query
        name: date_end
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.DailyTraffic'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: Ocserv Users Statistics
      tags:
      - Report
  /reports/total-bandwidth:
    get:
      consumes:
      - application/json
      description: Ocserv Users TotalBandwidth calculating
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: date_start
        in: query
        name: date_start
        required: true
        type: string
      - description: date_end
        in: query
        name: date_end
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repository.TotalBandwidths'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: Ocserv Users TotalBandwidth calculating
      tags:
      - Report
  /reports/users:
    get:
      consumes:
      - application/json
      description: Result of all user reports
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reports.OcservUserReportResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: Result of all user reports
      tags:
      - Report
...
  /system/ocserv-config:
    get:
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/runtime.ConfigResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/request.ErrorResponse'
      summary: Get structured Ocserv configuration
      tags:
      - System
    patch:
      consumes:
      - application/json
      description: Available only in normal/systemd mode; Docker deployments reject
        updates.
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Supported Ocserv configuration changes
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/runtime.OcservConfig'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/runtime.OcservConfig'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/request.ErrorResponse'
      summary: Update structured Ocserv configuration
      tags:
      - System
...
produces:
- application/json
swagger: "2.0"